8211975: move testlibrary/jdk/testlibrary/OptimalCapacity.java to top-level library Reviewed-by: igerasim
